- Created the "folder" and "Project" command. - Fixed warnings

+++ Engine/Lua/EngineLuaFunctionBindings.cpp
@@ -245,7 +245,7 @@ int LuaDrawSprite(lua_State* theState)
 	float ppu = LuaGetFloat(theState, 9, 1.f);
 
 	Vector2 dimensions = Vector2(width * SPRITE_DIMENSION * .5f, height * SPRITE_DIMENSION * .5f);
-	AABB2 uvs = g_theSpriteSheet->GetTexCoordsForSpriteIndexAndDimensions(spriteIndex, IntVector2(width, height));
+	AABB2 uvs = g_theSpriteSheet->GetTexCoordsForSpriteIndexAndDimensions(spriteIndex, IntVector2((int) width, (int) height));
 	Sprite spriteToDraw = Sprite(g_theSpriteSheet->m_texture, dimensions, uvs, ppu);
 
 	Renderer* r = Renderer::GetInstance();
